vaanikey_backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
audioread==3.1.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
decorator==5.2.1
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
filelock==3.20.3
fsspec==2026.1.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.24.7
HyperPyYAML==1.2.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.63.1
numpy==1.26.4
packaging==26.0
platformdirs==4.5.1
pooch==1.9.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
shellingham==1.5.4
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
speechbrain==1.0.3
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
tqdm==4.67.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3